requirements.txt
fastapi
pandas
uvicorn
requests
bs4
sqlalchemy
psycopg2-binary
passlib
python-jose
python-multipart